+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+Added
+~~~~~
+
+- `transcriptic.sampledata` module for enabling mocked Jupyter object exploration without establishing an explicit connection
+- example notebook for Jupyter object exploration
+
 Updated
 ~~~~~~~